feat(cli): susbtitute projectName construct stage-specific names (#880)
Updated all the templates to construct the name from the app name and stage

Example:
```ts
export const worker = await Astro("website", {
  name: `${app.name}-${app.stage}-website`,
});
```

Also removed the morph that would add `version: ` tag. Because DO migrations don't work in a preview worker, it's better to just deploy a whole separate worker.

A preview worker is only really helpful if you have a lot of static assets and want to avoid redundant uploads for each PR. But that should be done with knowledge of the drawbacks.